backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-limiter==3.5.0
PyJWT==2.8.0
qrcode==7.4.2
pillow==10.0.1
gunicorn==21.2.0 